@@ -1,6 +1,6 @@
 import 'package:places/data/model/category.dart';
 import 'package:places/data/model/geo_position.dart';
-import 'package:places/data/model/sight.dart';
+import 'package:places/data/model/place.dart';
 import 'dart:math';
 // ignore: import_of_legacy_library_into_null_safe
 import 'package:lorem_cutesum/lorem_cutesum.dart';
@@ -19,16 +19,18 @@ List<String> _randomUrls(int count) => [
       for (var i = 0; i < count; i++) Cutesum.randomImageUrl(),
     ];
 
-List<Sight> generateMocks(int count) {
+List<Place> generateMocks(int count) {
   return [
     for (var i = 0; i < count; i++)
-      Sight(
-        _randomWords(1, 4),
+      Place(
+        id: i + 1,
+        name: _randomWords(1, 4),
         geo: GeoPosition(
             (_rand.nextDouble() - .5) * 180, (_rand.nextDouble() - .5) * 360),
-        photoUrls: _randomUrls(_randomInt(1, 4)),
-        details: _randomWords(150, 300),
-        type: categories.keys.toList()[_rand.nextInt(categories.length)],
+        urls: _randomUrls(_randomInt(1, 4)),
+        description: _randomWords(150, 300),
+        category:
+            Category.byId.values.toList()[_rand.nextInt(Category.byId.length)],
       )
   ];
 }
